+++ changelog.md
@@ -1,3 +1,7 @@
+# 3.1.0
+- Updated to Geode v5.0.1
+- Now using event exports (thanks Jasmine)
+- Web request using Geode's web::WebRequest class can be overwritten (requires a cookie)
 # 3.0.0
 - Updated to Geode v5, new event system
 - Added `ServerUpdatingEvent`, a read-only Event which fires when the server updates; specifies what server is being switched to and by whom